requirements.txt
opencv-python
numpy
colorthief
Pillow
matplotlib